Make a new Search module, do a bit of cleanup
This change does a few cleanup-related things, including:

* Pulling the packages page and packages list into a new Search module, since what remains of those views now supports only search

* Removes a bunch of now-dead code from those components

* Fixes a bug with the progress bar by restricting it to origins-loading state (we aren’t yet set up to use it properly elsewhere)

* Routes links from the Explore view to /latest
