+++ components/builder-web/app/app.scss
@@ -43,7 +43,6 @@ $output-bourbon-deprecation-warnings: false !default;
 @import "origin/origin-page/origin-keys-tab/key-list/key-list.component";
 @import "origin/origin-page/origin-members-tab/origin-members-tab.component";
 @import "origin/origins-page/origins-page.component";
-@import "package-page/package-page";
 @import "package/package/package.component";
 @import "package/package-builds/package-builds.component";
 @import "package/package-latest/package-latest.component";
@@ -53,8 +52,8 @@ $output-bourbon-deprecation-warnings: false !default;
 @import "package/package-settings/package-settings.component";
 @import "package/package-sidebar/package-sidebar.component";
 @import "package/package-versions/package-versions.component";
-@import "packages-list/packages-list.component";
-@import "packages-page/packages-page";
+@import "search/results/results.component";
+@import "search/search/search.component";
 @import "project-create-page/project-create-page";
 @import "project-page/project-page";
 @import "projects-page/projects-page";
