--- components/builder-web/app/package-page/_package-page.scss
@@ -1,133 +1,0 @@
-// Copyright (c) 2016-2017 Chef Software Inc. and/or applicable contributors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-.project-header {
-  min-height: 2vh;
-  background-color: $hab-gray-light;
-  padding: 12px 0px 12px 25px;
-}
-
-.build-project-button {
-  margin-right: 25px;
-}
-
-.hab-package {
-  position: relative;
-
-  button.origin {
-    bottom: rem(12);
-    position: absolute;
-    right: rem(12);
-  }
-
-  .info, .builds {
-    @include outer-container;
-
-    .l {
-      @include span-columns(8);
-    }
-
-    .r {
-      @include span-columns(4);
-      @include omega;
-
-      border-left: 1px solid $hab-gray-light;
-      padding-left: $base-spacing;
-
-      li {
-        margin-bottom: $base-spacing * 1.5;
-      }
-    }
-  }
-
-  hab-icon[symbol="loading"] {
-    position: absolute;
-    right: rem(12);
-    top: rem(12);
-  }
-
-  &-info {
-    @include outer-container;
-    margin-bottom: $base-spacing;
-  }
-
-  &-list {
-    font-size: rem(14);
-    margin-bottom: $base-spacing;
-  }
-
-  &-deps {
-    h4 {
-      margin-top: $base-spacing;
-    }
-  }
-
-  &-latest-build {
-    div {
-      margin-bottom: $base-spacing;
-      font-size: rem(14);
-    }
-  }
-
-  &-manifest .manifest,
-  &-config pre {
-    background-color: darken($hab-off-white, 2);
-    border-radius: $global-radius;
-    padding: $small-spacing;
-  }
-
-  .builds {
-    font-family: $heading-font-family;
-    font-size: rem(12);
-    text-transform: uppercase;
-
-    a {
-      display: inline-block;
-      color: $dark-gray;
-      padding-bottom: 8px;
-
-      &:hover {
-        padding-bottom: 6px;
-        border-bottom: 2px solid $hab-green;
-      }
-    }
-  }
-}
-
-.hab-package-channels,
-.hab-package-manifest,
-.hab-package-config {
-  margin-bottom: rem(20);
-}
-
-.hab-package-manifest {
-
-  .manifest {
-    h1 {
-      font-size: rem(16);
-    }
-
-    h2,
-    h3,
-    p {
-      font-size: rem(14);
-    }
-  }
-
-  pre {
-    white-space: -moz-pre-wrap; // Mozilla, supported since 1999
-    white-space: pre-wrap;
-    word-wrap: break-word;
-  }
-}
