--- components/builder-web/app/packages-list/packages-list.component.html
@@ -1,20 +1,0 @@
-<ul class="hab-packages-list">
-    <div *ngIf="noPackages">
-        <p>
-        No packages found.
-        <span *ngIf="errorMessage">
-            Error: {{errorMessage}}
-        </span>
-        </p>
-    </div>
-    <div class="pkg-{{ layout }}">
-        <li class="hab-packages-package" *ngFor="let pkg of packages">
-            <a [routerLink]="routeFor(pkg)">
-                <div class="name" [class.large]="!pkg.version">
-                    <span>{{ packageString(pkg) }}</span>
-                </div>
-                <hab-icon symbol="chevron-right"></hab-icon>
-            </a>
-        </li>
-    </div>
-</ul>
