--- components/builder-web/app/packages-page/_packages-page.scss
@@ -1,99 +1,0 @@
-// Copyright (c) 2016-2017 Chef Software Inc. and/or applicable contributors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-.hab-packages {
-
-  .active {
-    @include row;
-
-    line-height: rem(42);
-    border-radius: 7px;
-    font-size: rem(14);
-    padding: 0 12px;
-    margin-bottom: 22px;
-
-    a {
-      text-decoration: underline;
-    }
-
-    &.dispatched, &.processing {
-      border: 1px solid $hab-orange;
-      color: $hab-orange;
-
-      a {
-        color: $hab-orange;
-      }
-    }
-
-    &.pending {
-      border: 1px solid $hab-blue;
-      color: $hab-blue;
-
-      a {
-        color: $hab-blue;
-      }
-    }
-  }
-
-  .builds {
-    font-family: $heading-font-family;
-    font-size: rem(12);
-    text-transform: uppercase;
-
-    a {
-      display: inline-block;
-      color: $dark-gray;
-      padding-bottom: 8px;
-
-      &:hover {
-        padding-bottom: 6px;
-        border-bottom: 2px solid $hab-green;
-      }
-    }
-  }
-
-  hab-icon[symbol="loading"] {
-    position: absolute;
-    right: 20px;
-    bottom: 20px;
-  }
-
-  &-list {
-    li a {
-      @extend .hab-item-list;
-    }
-  }
-
-  .build-history {
-    margin-top: 22px;
-    font-size: rem(14);
-  }
-
-  .page-body--sidebar {
-
-    h4 {
-      text-transform: uppercase;
-    }
-
-    .install-box {
-      border: 1px solid $medium-gray;
-      font-family: $monospace-font-family;
-      font-size: $monospace-font-size;
-      padding: 12px;
-      color: $dark-gray;
-      font-weight: 600;
-      overflow: auto;
-    }
-  }
-}
