--- components/builder-web/app/packages-page/packages-page.component.html
@@ -1,67 +1,0 @@
-<div class="hab-packages">
-    <div class="page-title">
-        <div *ngIf="showSearch">
-            <h2>Search Packages</h2>
-            <h4>
-                <span *ngIf="searchQuery || query">Search Results</span>
-            </h4>
-            <hab-icon symbol="loading" [class.spinning]="ui.loading"></hab-icon>
-        </div>
-        <div *ngIf="showBreadcrumbs">
-            <h2>
-                <hab-package-breadcrumbs
-         *ngIf="!searchQuery"
-         [ident]="packageParams()">
-                </hab-package-breadcrumbs>
-            </h2>
-            <h4>{{ subtitle }}</h4>
-        </div>
-    </div>
-    <div class="page-body" [class.has-sidebar]="iCanRequestABuild">
-        <div [class.page-body--main]="iCanRequestABuild">
-            <input *ngIf="showSearch"
-                type="search" autofocus
-                [formControl]="searchBox"
-                placeholder="Search Packages&hellip;">
-
-            <div class="active {{ activeBuild.state | lowercase }}" *ngIf="activeBuild">
-                A build is in progress.
-                <a [routerLink]="['/builds', activeBuild.id]">View streaming output</a>.
-            </div>
-
-            <hab-packages-list
-                 [noPackages]="(!ui.exists || packages.size === 0) && !ui.loading"
-                 [packages]="packages"
-                 [versions]="versions"
-                 [layout]="layout"
-                 [errorMessage]="ui.errorMessage"></hab-packages-list>
-
-             <div *ngIf="packages.size < totalCount">
-                 Showing {{packages.size}} of {{totalCount}} packages.
-                 <a href="#" (click)="fetchMorePackages()">
-                     Load
-                     {{(totalCount - packages.size) > perPage ? perPage : totalCount - packages.size }}
-                     more</a>.
-             </div>
-
-             <div class="build-history" *ngIf="showBuildHistoryLink">
-                <a [routerLink]="['/pkgs', origin, name, 'builds']">
-                    View full build history
-                    <hab-icon symbol="chevron-right"></hab-icon>
-                </a>
-            </div>
-        </div>
-        <div class="page-body--sidebar" *ngIf="iCanRequestABuild">
-            <h4>Build</h4>
-            <p>
-            <button class="button" (click)="requestNewBuild()" [disabled]="!!activeBuild">
-                Request new build
-            </button>
-            </p>
-            <h4>Install Latest Version</h4>
-            <div>
-                <pre class="install-box">hab install {{origin}}/{{name}}</pre>
-            </div>
-        </div>
-    </div>
-</div>
